requirements/mminstall.txt
mmcv-full>=1.3.1,<=1.4.0
mmdet>=2.11.0,<=2.12.0
